\name{fields-stuff}
\alias{fields.diagonalize}
\alias{fields.duplicated.matrix}
\alias{fields.mkpoly}
\title{Fields supporting functions}
\description{
Some supporting functions that are internal to fields top level methods. 
}
\usage{

}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{}{}

}

\details{

}
\author{Doug Nychka}
\seealso{Krig, Tps, as.image}
}
\keyword{spatial}% at least one, from doc/KEYWORDS
