\name{rat.diet}
\alias{rat.diet}
\title{
  Experiment studying an appetite supressant in rats.
}
\description{
The `rat.diet' data frame has 39 rows and 3 columns.
These are data from a study of an appetite supressant given to young rats.
The suppressant was removed from the treatment group at around 60 days.
The responses are the median food intake and  each group had approximately 
10 animals. 
}
\usage{
data(rat.diet)
}
\format{
This data frame contains the following columns:

\describe{
\item{t}{
Time in days}
\item{con}{
Median food intake of the control group }
\item{trt}{
Median food intake of the treatment group}
}

}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
