% fields, Tools for spatial data
% Copyright 2015, Institute for Mathematics Applied Geosciences
% University Corporation for Atmospheric Research
% Licensed under the GPL -- www.gpl.org/licenses/gpl.html

\name{lennon}
\alias{lennon}
\title{
  Gray image of John Lennon.
}
\description{
A 256X256 image of John Lennon.
Try:

\code{image(lennon, col=grey(seq(0,1,,256)) )}

}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
