\name{US.dat}
\alias{US.dat}
\title{
Outline of coterminous US and states.
}
\description{
This data set is used by the fields function US to draw a map. It is the
medium resolution outline that is produced by drawing the US from the maps
package.
}
\keyword{datasets}
% docclass is data
% Converted by Sd2Rd version 1.21.
